requirements.txt
Django==1.9.5
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
backports.ssl-match-hostname==3.5.0.1
django-jinja==2.1.2
docutils==0.12
lockfile==0.12.2
pymongo==3.2.2
python-daemon==2.1.1
requests==2.9.1
six==1.10.0
slackclient==1.0.0
websocket-client==0.37.0
wsgiref==0.1.2
